+++ src/io.h
@@ -34,8 +34,8 @@ struct read_info{
 	char* md;
 	//char* xp;
 	float mapq;
-	float prob;
-	float bar_prob; 
+	double prob;
+	double bar_prob; 
 	//float* priors;
 	//float* identity;
 	//int read_start;
